make grpcErr work for either plugin type
Extract a better function name and make the errors generic for different
plugin types.
